Rename copy-button.js to clipboard.js
The module is the site's one clipboard writer, shared by the heading
copy-URL button, the code block copy button and the "Copy as markdown"
page action. The old name read as though it belonged to a single button.

Splitting the write out from the button feedback was the alternative.
Every consumer wants both halves, so a split would add an import to each
call site and buy nothing until something needs to write to the clipboard
without a button.

`announce` and `REVERT_MS` were exported with no importer, so they go.

The two clipboard tests move to copy-button.spec.ts. They assert what the
copy button puts on the clipboard, which is not what llm-endpoints.spec.ts
is about; the whole suite runs against the same `astro preview` server, so
the built .md endpoints are reachable from either file.
